[+] Factorize handle search param in a service
